build: Rename socketMiddleware to socket-middleware
